requirement.txt
streamlit
pandas
python-dotenv
breeze-connect
python-telegram-bot
kiteconnect
openai
redis
